requirements.txt
aiohttp==3.5.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-dysmsapi==1.0.0
apispec==2.0.2
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-apispec==0.8.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.20.0
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
httplib2==0.13.0
idna==2.8
idna-ssl==1.1.0
inapppy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
multidict==4.5.2
numpy==1.16.4
oauth2client==3.0.0
paginate==0.5.6
paginate-sqlalchemy==0.3.0
pandas==0.24.2
passlib==1.7.1
pendulum==2.0.5
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodome==3.8.2
pycryptodomex==3.7.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
pytzdata==2019.2
qiniu==7.2.4
redis==3.2.1
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.5
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
webargs==5.3.1
pillow==6.2.1
Flask-Script==2.0.6
Flask-Migrate==2.2.1
flasgger==0.9.3
kubernetes==1.12.0
